fix: There is a bug in core Python (see: https://bugs.python.org/issue44831).
datetime.datetime.now and
datetime.datetime.fromtimestamp round
time differently. Standard library's logging
uses time.time to register time and while this
is compared to datetime.datetime.now, there
may be cases where the log record is in the
future. To fix this, time.time is used everywhere
where the time in log records are compared
to current time.