fix: An issue in registering and cond parsing.
There was an issue that if parsing a condition fails, the task was
already registered. However, the name setting and registering
cannot be done after setting conditions as those might require
a name for the task.

Separater name setting and registering. Now registering is done
at the very end of the initiation.